fixed bug type size allRAM, freeRAM reduction removed ~64GB max value, all memory available on my server

+++ src/Main_Window.cpp
@@ -3744,7 +3744,7 @@ void Main_Window::on_CH_Remove_RAM_Size_Limitation_stateChanged(int state)
     Update_RAM_Size_ComboBox(32768);
   }
   else {
-    int allRAM = 0, freeRAM = 0;
+    size_t allRAM = 0, freeRAM = 0;
     System_Info::Get_Free_Memory_Size(allRAM, freeRAM);
 
     if (allRAM < ui.Memory_Size->value())
@@ -3760,7 +3760,7 @@ void Main_Window::on_CH_Remove_RAM_Size_Limitation_stateChanged(int state)
 
 void Main_Window::on_TB_Update_Available_RAM_Size_clicked()
 {
-  int allRAM = 0, freeRAM = 0;
+  size_t allRAM = 0, freeRAM = 0;
   System_Info::Get_Free_Memory_Size(allRAM, freeRAM);
   ui.TB_Update_Available_RAM_Size->setText(
       tr("Free memory: %1 MB").arg(freeRAM));
